+++ src/FileCompiler.php
@@ -2,6 +2,7 @@
 
 namespace Surgiie\Blade;
 
+use Illuminate\View\AnonymousComponent as BladeAnonymousComponent;
 use Illuminate\View\Compilers\BladeCompiler;
 
 class FileCompiler extends BladeCompiler
@@ -14,7 +15,13 @@ class FileCompiler extends BladeCompiler
     protected $formatDirectives = [
         'foreach',
         'endforeach',
+        'component',
+        'endcomponent',
         'empty',
+        'slot',
+        'endslot',
+        'php',
+        'endphp',
         'if',
         'elseif',
         'else',
@@ -35,7 +42,9 @@ class FileCompiler extends BladeCompiler
      * Compile and rendered php code can leave behind unwanted spacing which
      * can be problematic for files where spacing has semantical meaning.
      * This function compiles known blade directives so that they are shifted
-     * to the start of the line should they have leading whitespace.
+     * to the start of the line should they have leading whitespace. This
+     * keeps the compiled content from shifting spaces and ending up in
+     * places where it shouldnt be.
      *
      * @see https://www.php.net/manual/en/language.basic-syntax.phptags.php
      */
@@ -43,9 +52,33 @@ class FileCompiler extends BladeCompiler
     {
         $keywords = implode('|', $this->formatDirectives);
 
-        $value = preg_replace("/\\s+\@($keywords)/", "\n@$1", $value);
+        // move all @ directives that are spaced/tabbed in to the start of the
+        // file this helps preserve the location of the content after compile
+        $value = preg_replace("/\\s+\@($keywords)/", PHP_EOL.'@$1', $value);
+        // add new line @endcomponent to ensure the next line isnt merged to the last line of component file.
+        $value = preg_replace('/@(endcomponent)/', '@$1'.PHP_EOL, $value);
+        // and a new line after @endComponentClass so the next line doesnt get merged to end of component file either.
+        $value = preg_replace('/@(endComponentClass)(.*)/', '@$1'.PHP_EOL, $value);
 
         return parent::compileStatements($value);
+    }
+
+    /**
+     * Compile a class component opening.
+     *
+     * @param  string  $component
+     * @param  string  $alias
+     * @param  string  $data
+     * @param  string  $hash
+     * @return string
+     */
+    public static function compileClassComponentOpening(string $component, string $alias, string $data, string $hash)
+    {
+        if ($component == BladeAnonymousComponent::class) {
+            $component = AnonymousComponent::class;
+        }
+
+        return parent::compileClassComponentOpening($component, $alias, $data, $hash);
     }
 
     /**
@@ -59,7 +92,23 @@ class FileCompiler extends BladeCompiler
      */
     public function isExpired($path)
     {
-        // ensures that compiler compiles the file always
         return true;
+    }
+
+    /**
+     * Compile the component tags.
+     *
+     * @param  string  $value
+     * @return string
+     */
+    protected function compileComponentTags($value)
+    {
+        if (! $this->compilesComponentTags) {
+            return $value;
+        }
+
+        return (new ComponentTagCompiler(
+            $this->classComponentAliases, $this->classComponentNamespaces, $this
+        ))->compile($value);
     }
 }