+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Release Notes
 
-## [Unreleased](https://github.com/surgiie/blade/compare/v0.1.0...master)
+## [Unreleased](https://github.com/surgiie/blade/compare/v0.2.0...master)
+## [v0.2.0] - 2022-10-26
+
+### Added
+
+- `@php`,`@component` to the list of directives that are shifted to start of line by @surgiie in https://github.com/surgiie/blade/pulls/1
+- `@endcomponent` gets a new line implictly added to avoid merging with next line on files by @surgiie in https://github.com/surgiie/blade/pulls/1
+- `<x-*>` blade component support by @surgiie in https://github.com/surgiie/blade/pulls/1
 
 ## [v0.1.0] - 2022-10-20
 
-Initial Release 
+Initial Release